pyproject.toml
[project]
name = "clinical-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
